backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic>=0.42.0
openai>=1.59.0
google-genai>=1.0.0
python-dotenv==1.0.1
pydantic>=2.10.0
httpx>=0.28.0
asyncpg>=0.30.0
